Extract ApplicationBar
